Improve parsing of escape sequences
Overhauled the code which parses escape sequences so that it properly
handles argument lists. Which gives a more accurate translation of VT100
escape sequences into generic curses commands. The set of implemented
escape sequences is still incomplete, but what is there is now more
correct and also scales much better.