Improve BDOS logging and fix a FCB bug
Significantly expand the logging of BDOS calls to make it easier to
follow what a CP/M program is doing. Such that after the program completes
the resultant log file can be grepped to make the sequence of BDOS calls
much easier to understand.

Also fix a long-standing issue with the FCB structure; this wouldn't have
caused any functionality difference but caused FCB-related logging to be
inaccurate.

Plus the usual grab-bag of minor misc improvements.