Improve terminal emulation and add documentation
The terminal emulation library has been overhauled in order to make it
easier to add new definitions in the future, fix a couple of issues with
the VT100 emulation, and several fixes for the Televideo emulation.

Televideo emulation is still quite spotty; it now is sufficient for some
CP/M binaries (e.g. WordStar and Turbo Pascal) but is still buggy on
others (e.g. MultiPlan).

Add a new status document to describe how well various aspects are
working.