Add support for configurable key mappings
Key mappings are now configurable. This is controlled by the new --keymap
option, and the default value for this is the new wordstar.keys file.

This allows ncurses keys (e.g. left arrow) to be mapped to particular
"raw" keys (such as ^S) to better support programs such as WordStar or
Turbo Pascal.