Add an option to disable protection of the warm start vector
Some programs such as SuperCalc modify the warm start vector. ZCPM
normally protects against this as modification of low memory can be a
symptom of an errant binary. This check can now be disabled via a new
command-line option.

Similarly, an option has been added for toggling protection of the BDOS
jump vector.