Use fmtlib instead of boost::format for formatted text
std::format is still not yet well supported by current compilers. Moving
to fmtlib is a good interim step because it is faster than boost::format
and its usage is quite close to that of std::format.

The long-term goal is to use std::format once it is available in both
current gcc and current clang.